LPS-62127 Rename to DLMimeTypeDisplayContext
